refactor(ui): factor Pak/IoStore tabs
