+++ aegis/ui/widgets/pak_iostore/__init__.py
@@ -1,8 +1,12 @@
 from .panel import PakIoStorePanel
+from .single_tab import SingleFileTab
+from .directory_tab import DirectoryTab
 from .utils import build_iostore_cmd, build_unrealpak_cmd, dest_for
 
 __all__ = [
     "PakIoStorePanel",
+    "SingleFileTab",
+    "DirectoryTab",
     "build_iostore_cmd",
     "build_unrealpak_cmd",
     "dest_for",
