+++ aegis/ui/widgets/pak_iostore/panel.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from pathlib import Path
-from typing import List, Optional
+from typing import Optional
 
 from PySide6.QtWidgets import (
     QCheckBox,
@@ -13,14 +12,13 @@ from PySide6.QtWidgets import (
     QPlainTextEdit,
     QPushButton,
     QTabWidget,
-    QTableWidget,
-    QTableWidgetItem,
     QVBoxLayout,
     QWidget,
 )
 
-from .worker import PakWorker, Task
-from .utils import build_iostore_cmd, build_unrealpak_cmd, dest_for
+from .directory_tab import DirectoryTab
+from .single_tab import SingleFileTab
+from .worker import PakWorker
 
 
 class PakIoStorePanel(QWidget):
@@ -38,14 +36,25 @@ class PakIoStorePanel(QWidget):
             self.preview.setPlainText,
             lambda: self.unrealpak_edit.text(),
             lambda: self.iostore_edit.text(),
-            lambda: self.pak_filter.text().strip(),
-            lambda: self.dir_unzip_extract.isChecked(),
-            self._set_row_status,
+            lambda: self.single_tab.pak_filter.text().strip(),
+            lambda: self.dir_tab.dir_unzip_extract.isChecked(),
+            self.dir_tab.set_row_status,
             self._on_worker_state,
+        )
+        self.single_tab.set_worker(
+            self.worker,
+            lambda: self.unrealpak_edit.text(),
+            lambda: self.iostore_edit.text(),
+        )
+        self.dir_tab.set_worker(
+            self.worker,
+            lambda: self.unrealpak_edit.text(),
+            lambda: self.iostore_edit.text(),
+            lambda: self.single_tab.pak_filter.text().strip(),
         )
         self._on_worker_state()
 
-    # ----- UI -----------------------------------------------------------
+    # ----- UI ---------------------------------------------------------
     def _build_ui(self) -> None:
         root = QVBoxLayout(self)
 
@@ -69,8 +78,10 @@ class PakIoStorePanel(QWidget):
         root.addWidget(tools_box)
 
         self.mode_tabs = QTabWidget()
-        self._build_single_tab()
-        self._build_dir_tab()
+        self.single_tab = SingleFileTab(self._log)
+        self.dir_tab = DirectoryTab(self._log)
+        self.mode_tabs.addTab(self.single_tab, "Single File")
+        self.mode_tabs.addTab(self.dir_tab, "Directory")
         root.addWidget(self.mode_tabs)
 
         self.eula_chk = QCheckBox("I certify I own these files")
@@ -86,181 +97,13 @@ class PakIoStorePanel(QWidget):
         self.log.setObjectName("pak_log")
         root.addWidget(self.log)
 
-    def _build_single_tab(self) -> None:
-        tab = QWidget()
-        layout = QVBoxLayout(tab)
-
-        self.pak_file = QLineEdit()
-        self.pak_browse = QPushButton("Browse")
-        row_pak = QHBoxLayout()
-        row_pak.addWidget(QLabel(".pak"))
-        row_pak.addWidget(self.pak_file, 1)
-        row_pak.addWidget(self.pak_browse)
-        layout.addLayout(row_pak)
-
-        self.pak_filter = QLineEdit()
-        self.pak_filter.setObjectName("pak_filter_edit")
-        row_filter = QHBoxLayout()
-        row_filter.addWidget(QLabel("Filter"))
-        row_filter.addWidget(self.pak_filter, 1)
-        layout.addLayout(row_filter)
-
-        self.utoc_file = QLineEdit()
-        self.utoc_browse = QPushButton("Browse")
-        row_utoc = QHBoxLayout()
-        row_utoc.addWidget(QLabel(".utoc"))
-        row_utoc.addWidget(self.utoc_file, 1)
-        row_utoc.addWidget(self.utoc_browse)
-        layout.addLayout(row_utoc)
-
-        self.obb_file = QLineEdit()
-        self.obb_browse = QPushButton("Browse")
-        row_obb = QHBoxLayout()
-        row_obb.addWidget(QLabel(".obb"))
-        row_obb.addWidget(self.obb_file, 1)
-        row_obb.addWidget(self.obb_browse)
-        layout.addLayout(row_obb)
-
-        self.aab_file = QLineEdit()
-        self.aab_browse = QPushButton("Browse")
-        row_aab = QHBoxLayout()
-        row_aab.addWidget(QLabel(".aab"))
-        row_aab.addWidget(self.aab_file, 1)
-        row_aab.addWidget(self.aab_browse)
-        layout.addLayout(row_aab)
-
-        btn_row = QHBoxLayout()
-        self.single_list_btn = QPushButton("List")
-        self.single_search_btn = QPushButton("Search")
-        self.single_extract_btn = QPushButton("Extract")
-        self.single_validate_btn = QPushButton("Validate")
-        btn_row.addWidget(self.single_list_btn)
-        btn_row.addWidget(self.single_search_btn)
-        btn_row.addWidget(self.single_extract_btn)
-        btn_row.addWidget(self.single_validate_btn)
-        btn_row.addStretch(1)
-        layout.addLayout(btn_row)
-
-        self.mode_tabs.addTab(tab, "Single File")
-
-    def _build_dir_tab(self) -> None:
-        tab = QWidget()
-        layout = QVBoxLayout(tab)
-
-        dir_row = QHBoxLayout()
-        self.dir_root = QLineEdit()
-        self.dir_browse = QPushButton("Browse")
-        dir_row.addWidget(QLabel("Root directory"))
-        dir_row.addWidget(self.dir_root, 1)
-        dir_row.addWidget(self.dir_browse)
-        layout.addLayout(dir_row)
-
-        self.dir_recurse = QCheckBox("Recurse subdirectories")
-        self.dir_recurse.setChecked(True)
-        self.dir_unzip = QCheckBox("Also unzip .obb/.aab files")
-        self.dir_unzip.setChecked(True)
-        self.dir_unzip_extract = QCheckBox(
-            "After unzip, also extract any .pak found inside"
-        )
-        self.dir_unzip_extract.setChecked(True)
-        layout.addWidget(self.dir_recurse)
-        layout.addWidget(self.dir_unzip)
-        layout.addWidget(self.dir_unzip_extract)
-
-        self.dir_scan_btn = QPushButton("Scan")
-        layout.addWidget(self.dir_scan_btn)
-
-        self.dir_search = QLineEdit()
-        self.dir_search.setPlaceholderText("Search...")
-        layout.addWidget(self.dir_search)
-
-        self.table = QTableWidget(0, 4)
-        self.table.setHorizontalHeaderLabels(["Type", "Path", "Size", "Status"])
-        self.table.setObjectName("pak_table")
-        self.table.horizontalHeader().setStretchLastSection(True)
-        layout.addWidget(self.table)
-
-        btn_row = QHBoxLayout()
-        self.dir_extract_sel = QPushButton("Extract Selected")
-        self.dir_extract_all = QPushButton("Extract All")
-        self.dir_validate_sel = QPushButton("Validate Selected")
-        self.dir_validate_all = QPushButton("Validate All")
-        btn_row.addWidget(self.dir_extract_sel)
-        btn_row.addWidget(self.dir_extract_all)
-        btn_row.addWidget(self.dir_validate_sel)
-        btn_row.addWidget(self.dir_validate_all)
-        btn_row.addStretch(1)
-        layout.addLayout(btn_row)
-
-        cmp_box = QGroupBox("Compare Builds")
-        cmp_layout = QVBoxLayout()
-        row_a = QHBoxLayout()
-        self.compare_a = QLineEdit()
-        self.compare_a_browse = QPushButton("Browse")
-        row_a.addWidget(QLabel("Dir A"))
-        row_a.addWidget(self.compare_a, 1)
-        row_a.addWidget(self.compare_a_browse)
-        cmp_layout.addLayout(row_a)
-        row_b = QHBoxLayout()
-        self.compare_b = QLineEdit()
-        self.compare_b_browse = QPushButton("Browse")
-        row_b.addWidget(QLabel("Dir B"))
-        row_b.addWidget(self.compare_b, 1)
-        row_b.addWidget(self.compare_b_browse)
-        cmp_layout.addLayout(row_b)
-        self.compare_btn = QPushButton("Compare Builds")
-        cmp_layout.addWidget(self.compare_btn)
-        cmp_box.setLayout(cmp_layout)
-        layout.addWidget(cmp_box)
-
-        self.status_label = QLabel("Idle")
-        layout.addWidget(self.status_label)
-
-        self.mode_tabs.addTab(tab, "Directory")
-
-    # ----- Connections --------------------------------------------------
+    # ----- Connections -----------------------------------------------
     def _connect(self) -> None:
         self.unrealpak_browse.clicked.connect(self._pick_unrealpak)
         self.iostore_browse.clicked.connect(self._pick_iostore)
+        self.eula_chk.toggled.connect(self._update_enabled)
 
-        self.pak_browse.clicked.connect(
-            lambda: self._pick_file(self.pak_file, "Pak Files (*.pak)")
-        )
-        self.utoc_browse.clicked.connect(
-            lambda: self._pick_file(self.utoc_file, "IoStore Files (*.utoc)")
-        )
-        self.obb_browse.clicked.connect(
-            lambda: self._pick_file(self.obb_file, "OBB Files (*.obb)")
-        )
-        self.aab_browse.clicked.connect(
-            lambda: self._pick_file(self.aab_file, "AAB Files (*.aab)")
-        )
-        self.single_list_btn.clicked.connect(self._single_list)
-        self.single_search_btn.clicked.connect(self._single_search)
-        self.single_extract_btn.clicked.connect(self._single_extract)
-        self.single_validate_btn.clicked.connect(self._single_validate)
-
-        self.dir_browse.clicked.connect(self._pick_dir)
-        self.dir_scan_btn.clicked.connect(self.scan_directory)
-        self.dir_search.textChanged.connect(self._filter_table)
-        self.dir_extract_sel.clicked.connect(self._dir_extract_selected)
-        self.dir_extract_all.clicked.connect(self._dir_extract_all)
-        self.dir_validate_sel.clicked.connect(self._dir_validate_selected)
-        self.dir_validate_all.clicked.connect(self._dir_validate_all)
-        self.compare_a_browse.clicked.connect(
-            lambda: self._pick_dir_into(self.compare_a)
-        )
-        self.compare_b_browse.clicked.connect(
-            lambda: self._pick_dir_into(self.compare_b)
-        )
-        self.compare_btn.clicked.connect(self._compare_builds)
-
-        self.eula_chk.toggled.connect(self._update_extract_enabled)
-        self.table.itemSelectionChanged.connect(self._update_extract_enabled)
-
-        self._update_extract_enabled()
-
-    # ----- Helpers ------------------------------------------------------
+    # ----- Helpers ---------------------------------------------------
     def _pick_unrealpak(self) -> None:
         path, _ = QFileDialog.getOpenFileName(self, "UnrealPak", "", "Executables (*)")
         if path:
@@ -273,213 +116,16 @@ class PakIoStorePanel(QWidget):
         if path:
             self.iostore_edit.setText(path)
 
-    def _pick_file(self, edit: QLineEdit, filt: str) -> None:
-        path, _ = QFileDialog.getOpenFileName(self, "Choose file", "", filt)
-        if path:
-            edit.setText(path)
-        self._update_extract_enabled()
-
-    def _pick_dir(self) -> None:
-        path = QFileDialog.getExistingDirectory(self, "Choose directory")
-        if path:
-            self.dir_root.setText(path)
-
     def _log(self, text: str) -> None:
         self.log.appendPlainText(text)
 
-    def _set_row_status(self, row: int, status: str) -> None:
-        self.table.setItem(row, 3, QTableWidgetItem(status))
-
-    def _filter_table(self, text: str) -> None:
-        text = text.lower()
-        for row in range(self.table.rowCount()):
-            item = self.table.item(row, 1)
-            self.table.setRowHidden(row, text not in item.text().lower())
-
-    def _pick_dir_into(self, edit: QLineEdit) -> None:
-        path = QFileDialog.getExistingDirectory(self, "Choose directory")
-        if path:
-            edit.setText(path)
-
-    def _compare_builds(self) -> None:
-        a = Path(self.compare_a.text())
-        b = Path(self.compare_b.text())
-        if not a.is_dir() or not b.is_dir():
-            self._log("Invalid compare directories")
-            return
-        task = Task("cmp", "compare", a, b, None)
-        self.worker.enqueue([task])
-
     def _on_worker_state(self) -> None:
         running, pending = self.worker.status()
-        if running or pending:
-            self.status_label.setText(f"Running {running + pending} tasks...")
-        else:
-            self.status_label.setText("Idle")
-        self._update_extract_enabled()
+        self.dir_tab.update_status(running, pending)
+        self._update_enabled()
 
-    def _update_extract_enabled(self) -> None:
+    def _update_enabled(self) -> None:
         busy = self.worker.is_busy()
-        enable = self.eula_chk.isChecked() and not busy
-        self.single_list_btn.setEnabled(not busy)
-        self.single_search_btn.setEnabled(not busy)
-        self.single_extract_btn.setEnabled(enable)
-        self.single_validate_btn.setEnabled(enable)
-        self.dir_scan_btn.setEnabled(not busy)
-        has_rows = self.table.rowCount() > 0
-        has_sel = len(self.table.selectedIndexes()) > 0
-        self.dir_extract_sel.setEnabled(enable and has_sel)
-        self.dir_extract_all.setEnabled(enable and has_rows)
-        self.dir_validate_sel.setEnabled(enable and has_sel)
-        self.dir_validate_all.setEnabled(enable and has_rows)
-        self.compare_btn.setEnabled(not busy)
-
-    # ----- Scan ---------------------------------------------------------
-    def scan_directory(self) -> None:
-        root = Path(self.dir_root.text())
-        if not root.is_dir():
-            self._log(f"Invalid directory: {root}")
-            return
-        self.table.setRowCount(0)
-        paths = root.rglob("*") if self.dir_recurse.isChecked() else root.glob("*")
-        for p in paths:
-            if not p.is_file():
-                continue
-            ext = p.suffix.lower()
-            if ext == ".pak":
-                name = p.name.lower()
-                if "dlc" in name or name.endswith("_p.pak"):
-                    kind = "PAK-DLC"
-                else:
-                    kind = "PAK"
-            elif ext == ".utoc":
-                kind = "UTOC"
-            elif ext in {".obb", ".aab"} and self.dir_unzip.isChecked():
-                kind = "ZIP"
-            else:
-                continue
-            row = self.table.rowCount()
-            self.table.insertRow(row)
-            self.table.setItem(row, 0, QTableWidgetItem(kind))
-            self.table.setItem(row, 1, QTableWidgetItem(str(p)))
-            self.table.setItem(row, 2, QTableWidgetItem(str(p.stat().st_size)))
-            self.table.setItem(row, 3, QTableWidgetItem("Pending"))
-        self._update_extract_enabled()
-
-    # ----- Task queue ---------------------------------------------------
-    def _dir_extract_selected(self) -> None:
-        rows = {i.row() for i in self.table.selectedIndexes()}
-        self._enqueue_rows(sorted(rows), "extract")
-
-    def _dir_extract_all(self) -> None:
-        self._enqueue_rows(list(range(self.table.rowCount())), "extract")
-
-    def _dir_validate_selected(self) -> None:
-        rows = {i.row() for i in self.table.selectedIndexes()}
-        self._enqueue_rows(sorted(rows), "validate")
-
-    def _dir_validate_all(self) -> None:
-        self._enqueue_rows(list(range(self.table.rowCount())), "validate")
-
-    def _enqueue_rows(self, rows: List[int], action: str) -> None:
-        tasks: List[Task] = []
-        for row in rows:
-            path = Path(self.table.item(row, 1).text())
-            ttype = self.table.item(row, 0).text()
-            if ttype.startswith("PAK"):
-                dest = None if action != "extract" else dest_for(path, "_extracted")
-                argv = build_unrealpak_cmd(
-                    self.unrealpak_edit.text(),
-                    path,
-                    action if action != "search" else "list",
-                    dest,
-                    self.pak_filter.text().strip() or None,
-                )
-                tasks.append(Task("proc", action, path, dest, row, argv))
-            elif ttype.startswith("UTOC"):
-                dest = None if action != "extract" else dest_for(path, "_extracted")
-                argv = build_iostore_cmd(
-                    self.iostore_edit.text(),
-                    path,
-                    action if action != "search" else "list",
-                    dest,
-                )
-                tasks.append(Task("proc", action, path, dest, row, argv))
-            elif ttype == "ZIP":
-                dest = None if action != "extract" else dest_for(path, "_unzipped")
-                tasks.append(Task("zip", action, path, dest, row))
-        self.worker.enqueue(tasks)
-
-    def _single_list(self) -> None:
-        task = self._single_task("list")
-        if task:
-            self.worker.enqueue([task])
-
-    def _single_search(self) -> None:
-        task = self._single_task("search")
-        if task:
-            self.worker.enqueue([task])
-
-    def _single_extract(self) -> None:
-        if not self.eula_chk.isChecked():
-            self._log("Ownership checkbox not ticked")
-            return
-        task = self._single_task("extract")
-        if task:
-            self.worker.enqueue([task])
-
-    def _single_validate(self) -> None:
-        task = self._single_task("validate")
-        if task:
-            self.worker.enqueue([task])
-
-    def _single_task(self, action: str) -> Optional[Task]:
-        mapping = [
-            (self.pak_file, "unrealpak", "_extracted"),
-            (self.utoc_file, "iostore", "_extracted"),
-            (self.obb_file, "zip", "_unzipped"),
-            (self.aab_file, "zip", "_unzipped"),
-        ]
-        for edit, kind, suf in mapping:
-            if edit.text():
-                path = Path(edit.text())
-                if not path.exists():
-                    self._log(f"Missing file: {path}")
-                    return None
-                if kind == "unrealpak":
-                    mode = (
-                        action if action in {"list", "extract", "validate"} else "list"
-                    )
-                    dest = (
-                        None
-                        if action in {"list", "search", "validate"}
-                        else dest_for(path, suf)
-                    )
-                    argv = build_unrealpak_cmd(
-                        self.unrealpak_edit.text(),
-                        path,
-                        mode,
-                        dest,
-                        self.pak_filter.text().strip() or None,
-                    )
-                    return Task("proc", action, path, dest, None, argv)
-                if kind == "iostore":
-                    mode = (
-                        action if action in {"list", "extract", "validate"} else "list"
-                    )
-                    dest = (
-                        None
-                        if action in {"list", "search", "validate"}
-                        else dest_for(path, suf)
-                    )
-                    argv = build_iostore_cmd(self.iostore_edit.text(), path, mode, dest)
-                    return Task("proc", action, path, dest, None, argv)
-                tkind = "zip"
-                dest = (
-                    None
-                    if action in {"list", "search", "validate"}
-                    else dest_for(path, suf)
-                )
-                return Task(tkind, action, path, dest, None)
-        self._log("No file selected")
-        return None
+        eula = self.eula_chk.isChecked()
+        self.single_tab.update_enabled(busy, eula)
+        self.dir_tab.update_enabled(busy, eula)
